+++ src/modules/auth/auth.module.ts
@@ -1,18 +1,24 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
+import { AUTH_TOKEN_AUTHENTICATOR, AuthSessionService, JwtAuthGuard, TokenService } from '@wlisfes/chat-web-base-schema/auth'
 import { TbAccountUser } from '@wlisfes/chat-web-base-schema/chat-web-account-mysql'
 import { AuthController } from '@/modules/auth/auth.controller'
 import { AuthService } from '@/modules/auth/auth.service'
-import { AuthSessionService } from '@/modules/auth/auth-session.service'
 import { CaptchaService } from '@/modules/auth/captcha.service'
-import { JwtAuthGuard } from '@/modules/auth/jwt-auth.guard'
 import { PasswordService } from '@/modules/auth/password.service'
-import { TokenService } from '@/modules/auth/token.service'
 
 @Module({
     imports: [TypeOrmModule.forFeature([TbAccountUser])],
     controllers: [AuthController],
-    providers: [AuthService, AuthSessionService, CaptchaService, JwtAuthGuard, PasswordService, TokenService],
+    providers: [
+        AuthService,
+        AuthSessionService,
+        CaptchaService,
+        JwtAuthGuard,
+        PasswordService,
+        TokenService,
+        { provide: AUTH_TOKEN_AUTHENTICATOR, useExisting: AuthService }
+    ],
     exports: [AuthService, JwtAuthGuard, PasswordService, TokenService]
 })
 export class AuthModule {}